Move appendStackTrace tests to a dedicated file
The heavy GC loop tests in error-gc-test.test.js exceed the per-test
timeout on slow debug ASAN runs regardless of this change, so the new
tests now live in their own file. Also pass symbolize=0 to the spawned
child so an ASan abort on the unfixed build fails the test quickly
instead of hitting the test timeout during report symbolization.
