+++ test/js/bun/util/error-gc-test.test.js
@@ -1,6 +1,6 @@
 import { expect, test } from "bun:test";
 import { readFileSync } from "fs";
-import { bunEnv, bunExe, tmpdirSync } from "harness";
+import { tmpdirSync } from "harness";
 import { join } from "path";
 // This test checks that printing stack traces increments and decrements
 // reference-counted strings
@@ -27,49 +27,6 @@ test("error gc test", () => {
     fn();
     Bun.gc(true);
   }
-});
-
-// Appending an error's stack trace to itself made Vector::appendVector read
-// from its own freed buffer once the append grew past the vector's capacity.
-// Malloc=1 routes WTF allocations through the system allocator so ASan builds
-// can see the use-after-free.
-test("Error.appendStackTrace with the same error as source and destination", async () => {
-  const code = `
-    function f(n) {
-      if (n > 0) return f(n - 1) + 1;
-      try {
-        null();
-      } catch (e) {
-        Error.appendStackTrace(e, e);
-      }
-      return 0;
-    }
-    f(64);
-    console.log("ok");
-  `;
-  await using proc = Bun.spawn({
-    cmd: [bunExe(), "-e", code],
-    env: { ...bunEnv, Malloc: "1" },
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-  const [stdout, , exitCode] = await Promise.all([proc.stdout.text(), proc.stderr.text(), proc.exited]);
-  expect(stdout).toBe("ok\n");
-  expect(exitCode).toBe(0);
-});
-
-test("Error.appendStackTrace moves the source stack trace into the destination", () => {
-  function inner() {
-    try {
-      null();
-    } catch (e) {
-      return e;
-    }
-  }
-  const src = inner();
-  const dst = new Error("dst");
-  Error.appendStackTrace(src, dst);
-  expect(dst.stack).toContain("inner");
 });
 
 test("error gc test #2", () => {
